+++ src/Driver/CloudflareDriver.php
@@ -32,6 +32,7 @@
 namespace GlpiPlugin\Domainmanager\Driver;
 
 use DateTimeImmutable;
+use GlpiPlugin\Domainmanager\Config\Config;
 use GlpiPlugin\Domainmanager\Contract\ConnectionTestableInterface;
 use GlpiPlugin\Domainmanager\Contract\DnsPipelineInterface;
 use GlpiPlugin\Domainmanager\Contract\DnsRecordCommentSyncInterface;
@@ -493,6 +494,7 @@ class CloudflareDriver implements RegistrarDriverInterface, DnsPipelineInterface
      */
     public function createRecord(string $domain, string $type, string $name, string $data, int $ttl): ZoneRecord
     {
+        Config::assertWritesAllowed();
         $type      = self::assertWritableType($type);
         $domain    = self::normalizeDomain($domain);
         $accountId = $this->requireAccountId();
@@ -538,6 +540,7 @@ class CloudflareDriver implements RegistrarDriverInterface, DnsPipelineInterface
      */
     public function updateRecord(string $domain, string $remoteId, string $type, string $name, string $data, int $ttl): ZoneRecord
     {
+        Config::assertWritesAllowed();
         $type      = self::assertWritableType($type);
         $domain    = self::normalizeDomain($domain);
         $accountId = $this->requireAccountId();
@@ -574,6 +577,7 @@ class CloudflareDriver implements RegistrarDriverInterface, DnsPipelineInterface
      */
     public function deleteRecord(string $domain, string $remoteId): void
     {
+        Config::assertWritesAllowed();
         $domain    = self::normalizeDomain($domain);
         $accountId = $this->requireAccountId();
         $zoneId    = $this->findZone($domain, $accountId);
@@ -700,6 +704,7 @@ class CloudflareDriver implements RegistrarDriverInterface, DnsPipelineInterface
      */
     public function setProxied(string $domain, string $remoteId, bool $proxied): ZoneRecord
     {
+        Config::assertWritesAllowed();
         $domain    = self::normalizeDomain($domain);
         $accountId = $this->requireAccountId();
         $zoneId    = $this->findZone($domain, $accountId);
@@ -728,6 +733,7 @@ class CloudflareDriver implements RegistrarDriverInterface, DnsPipelineInterface
      */
     public function pushComment(string $domain, string $remoteId, string $comment): void
     {
+        Config::assertWritesAllowed();
         $domain    = self::normalizeDomain($domain);
         $accountId = $this->requireAccountId();
         $zoneId    = $this->findZone($domain, $accountId);